Fix PVS-Studio warning: implicit int->qreal conversions cause precision loss
V601: The bool type is implicitly cast to the integer type
V636: Expression was implicitly cast from 'int' type to 'double' type

CCBUG:393431

+++ libs/ui/widgets/kis_preset_live_preview_view.cpp
@@ -198,7 +198,7 @@ void KisPresetLivePreviewView::setupAndPaintStroke()
     double textureScale = settings->getDouble("Texture/Pattern/Scale");
     if ( textureOffsetX*textureScale> maxTextureSize || textureOffsetY*textureScale > maxTextureSize) {
         int maxSize = qMax(textureOffsetX, textureOffsetY);
-        double result = maxTextureSize/maxSize;
+        double result = qreal(maxTextureSize) / maxSize;
         settings->setProperty("Texture/Pattern/Scale", result);
     }
     if (m_currentPreset->paintOp().id() == "spraybrush") {
